CR + testing changes
