+++ monkey/infection_monkey/exploit/HostExploiter.py
@@ -35,13 +35,10 @@ class HostExploiter(Plugin):
     # Usual values are 'vulnerability' or 'brute_force'
     EXPLOIT_TYPE = ExploitType.VULNERABILITY
 
-    # Specifies whether a machine, on which the exploit was successful, should be added to the set of exploited
-    # machines. This would then prevent any other exploits from being attempted on it.
-    # Sample use case - Zerologon exploiter:
-    #     Exploited machine gives us useful credentials which can be used, but machine isn't fully compromised by Zerologon
-    #     on its own. Some other exploiter using PTH needs to exploit it with the extracted credentials so that the monkey
-    #     can propagate to it.
-    SHOULD_ADD_MACHINE_TO_EXPLOITED_SET = True
+    # Determines if successful exploitation should stop further exploit attempts on that machine.
+    # Generally, should be True for RCE type exploiters and False if we don't expect the exploiter to run the monkey agent.
+    # Example: Zerologon steals credentials
+    RUNS_AGENT_ON_SUCCESS = True
 
     @property
     @abstractmethod
@@ -109,4 +106,5 @@ class HostExploiter(Plugin):
         :param cmd: String of executed command. e.g. 'echo Example'
         """
         powershell = True if "powershell" in cmd.lower() else False
-        self.exploit_info['executed_cmds'].append({'cmd': cmd, 'powershell': powershell})
+        self.exploit_info['executed_cmds'].append(
+            {'cmd': cmd, 'powershell': powershell})
